requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.6.1,<2.0.0
langchain==0.0.350
streamlit==1.28.2
python-dotenv==1.0.0
pydantic==1.10.12
requests==2.31.0
chromadb==0.4.18
langchain-text-splitters==0.0.1
